Split server-questions tests into a folder and drop its max-lines exemption
The admin-questions server test file had grown to 1,532 lines and was
grandfathered into biome's noExcessiveLinesPerFile exemption list. Split
it into test/lib/server-questions/ — grouped by what each part checks
(the question resource, question deletes, answers + ordering, the answer
edit/recalculate pages, and listing-question links) — with the shared
createQuestion/addAnswer helpers lifted into helpers.ts, so every file
sits under 400 lines, then removed the file's exemption entry.

Bringing the file back under jscpd's per-file scan cap surfaced a
repeated "reject a bad modifier_id on answer edit" block that had been
passing silently inside the monolith; it is now a single local helper
(expectModifierRejected). All original tests carry over unchanged and
pass; jscpd is back to 0% duplication.
